Add methods initiate() and process() to module Wallet
A call to initiate() either does nothing or generates a new transaction.
If a transaction is generated, inputs are built by selecting utxos from the wallet
and outputs are chosen at random. But it does not mutate the wallet.

The process() method modifies the wallet accordingly (if the given transaction consumes
utxos from the wallet).